Store dataset labels as int8 and cast at loss time
Agent-Logs-Url: https://github.com/TEAT-Unicaen/R.A.I.D.-RAM-Artificial-Inspection-Dump/sessions/4ced8cd4-801b-47ce-b9d0-8a4e6e7ffc29

+++ src/dumpManager/RamDumpDataset.py
@@ -62,7 +62,7 @@ class RamDumpDataset(Dataset):
         chunk = self.ram_data[data_start : data_start + self.chunk_size]
         
         x = torch.tensor(list(chunk), dtype=torch.long)
-        y = torch.zeros(self.chunk_size, dtype=torch.float)
+        y = torch.zeros(self.chunk_size, dtype=torch.int8)
 
         temp_idx = meta_idx
         while temp_idx < len(self.metadata):
